+++ docs/release-notes.md
@@ -1,6 +1,7 @@
 # Release notes
 ## Upcoming release
 * For players:
+  * Fixed cryptic error message when the game isn't installed correctly.
   * Fixed error when a mod makes invalid changes to an NPC schedule.
   * Fixed invalid NPC data propagated when a mod changes NPC dispositions.
   * Fixed `Display.RenderedWorld` event broken in SMAPI 2.9.1.